// citpwawcl/src/app/rapprova/models/save-rapprova-web.model.ts
interface DatiRapProva {
  idAllegato?: number;
  dataControllo: number;
  codiceImpianto: number;
  fkTipoDocumento: number;
  fOraArrivo: string;
  elencoApparecchiature: string;
  elencoCombustibili: string;
  fkIspezIspet: number;
}

interface Modulo {
  s8aNModuloTermico: string;
  s8bFumoMis1: number;
  s8bFumoMis2: number;
  s8bFumoMis3: number;
  s8cMarcaStrumMisura: string;
  s8cModelloStrumMisura: string;
  s8cMatricolaStrumMisura: string;
  s8dTempFluidoMandataC: number;
  s8dTempAriaC: number;
  s8dTempFumiC: number;
  s8dO2Perc: number;
  s8dCo2Perc: number;
  s8dCoFumiSecchiPpm: number;
  s8dNoMgKwH: number;
  s8eIndiceAria: number;
  s8eFumiSecchiNoAriaPpm: number;
  s8eQsPerc: number;
  s8eEtPerc: number;
  s8eRendCombPerc: number;
  s8eNoxMgKwH: number;
  s9aFlgMonossidoCarb: string;
  s9bFlgFumosita: string;
  s9cRendMinCombustPerc: number;
  s9cFlgRendCombustSuff: number;
  s9dOssidiAzotoLimMgKwH: number;
  s9dFlgOssidiAzoto: string;
  s9eFlgRispettoNormativa: number;
  s9eFlgNoRispetto7a: number;
  s9eFlgNoRispetto7b: number;
  s9eFlgNoRispetto9a: number;
  s9eFlgNoRispetto9b: number;
  s9eFlgNoRispetto9c: number;
  s9eFlgNoRispetto9d: number;
}

interface DatiRapProvaWebGt {
  s1cFlgReeInviato: number;
  s1cFlgReeBollino: number;
  s1cSiglaBollino: string;
  s1cNumBollino: number;
  s1eDtPrimaInstallazione: number;
  s1ePotFocolareKw: number;
  s1ePotUtileKw: number;
  s1lDenomDelegato: string;
  s1lFlgDelega: number;
  s2b1FlgTermoContab: number;
  s2b2FlgUni10200: number;
  s2fFlgTrattClimaNonRich: number;
  s2fFlgTrattAcsNonRich: number;
  s3aFlgLocaleIntIdoneo: number;
  s3bFlgGenExtIdoneo: number;
  s3cFlgVentilazSuff: number;
  s3dFlgEvacFumiIdoneo: number;
  s3eFlgCartelliPresenti: number;
  s3fFlgEstinzPresenti: number;
  s3gFlgInterrGenPresenti: number;
  s3hFlgRubinPresente: number;
  s3iFlgAssenzaPerdComb: number;
  s3jFlgTempAmbFunz: number;
  s3kFlgDm1121975: number;
  s4aFlgLibImpPresente: number;
  s4bFlgLibCompilato: number;
  s4cFlgConformitaPresente: number;
  s4dFlgLibUsoPresente: number;
  s4eFlgPraticaVvfPresente: number;
  s4fFlgPraticaInailPresente: number;
  s4gFlgDm121975: number;
  s4gMatricolaDm1121975: string;
  s5aFlgAdozioneValvoleTerm: number;
  s5aFlgIsolamenteRete: number;
  s5aFlgAdozSistTrattamH2o: number;
  s5aFlgSostituzSistRegolaz: number;
  s5bFlgNoIntervConv: number;
  s5bFlgRelazDettaglio: number;
  s5bFlgRelazDettaglioSucc: number;
  s5bFlgValutazNonEseguita: number;
  s5bMotivoRelazNonEseg: string;
  s5cFlgDimensCorretto: number;
  s5cFlgDimensNonControll: number;
  s5cFlgDimensRelazSucces: number;
  s1cDataRee: number;
  s5cFlgDimensNonCorretto: number;
  idDettIspezGt?: number;
  progressivo: number;
  dataInstall: number;
  s6dFlgEvacuFumi: string;
  s6iFlgTipoB: number;
  s6iFlgTipoC: number;
  s6jFkClassDpr66096: string;
  s6kPotTermFocolKw: number;
  s6kBruciatoreDaKw: number;
  s6kBruciatoreAKw: number;
  s6lPortataCombM3H: number;
  s6lPortataCombKgH: number; // Nel BE è string, ma probabilmente number è il tipo corretto
  s6lPotTermFocolKw: number;
  s7aFlgManutEffettuata: number;
  s7aDataUltimaManut: number;
  s7bFlgReePresente: number;
  s7bDataRee: number;
  s7bFlgOsservazioni: number;
  s7bFlgRaccomand: number;
  s7bFlgPrescr: number;
  moduli: Modulo[];
  /*s8aNModuloTermico: string[];
  s8bFumoMis1: number[];
  s8bFumoMis2: number[];
  s8bFumoMis3: number[];
  s8cMarcaStrumMisura: string[];
  s8cModelloStrumMisura: string[];
  s8cMatricolaStrumMisura: string[];
  s8dTempFluidoMandataC: number[];
  s8dTempAriaC: number[];
  s8dTempFumiC: number[];
  s8dO2Perc: number[];
  s8dCo2Perc: number[];
  s8dCoFumiSecchiPpm: number[];
  s8dNoMgKwH: number[];
  s8eIndiceAria: number[];
  s8eFumiSecchiNoAriaPpm: number[];
  s8eQsPerc: number[];
  s8eEtPerc: number[];
  s8eRendCombPerc: number[];
  s8eNoxMgKwH: number[];
  s9aFlgMonossidoCarb: string[];
  s9bFlgFumosita: string[];
  s9cRendMinCombustPerc: number[];
  s9cFlgRendCombustSuff: number[];
  s9dOssidiAzotoLimMgKwH: number[];
  s9dFlgOssidiAzoto: string[];
  s9eFlgRispettoNormativa: number[];
  s9eFlgNoRispetto7a: number[];
  s9eFlgNoRispetto7b: number[];
  s9eFlgNoRispetto9a: number[];
  s9eFlgNoRispetto9b: number[];
  s9eFlgNoRispetto9c: number[];
  s9eFlgNoRispetto9d: number[];*/
  // controlloweb: number;
  s7aFrequenzaManutAltro: string;
  s7aFkFrequenzaManut: number;
  fOsservazioni: string;
  fRaccomandazioni: string;
  fPrescrizioni: string;
}

interface Circuito {
  s8aNCircuito: string;
  s8bFlgProveRaffrescamento: number;
  s8bFlgProveRiscaldamento: number;
  s8cFlgFiltriPuliti: number;
  s8dFlgAssenzaPerditeGas: number;
  s8eMarcaStrumMisura: string;
  s8eModelloStrumMisura: string;
  s8eMatricolaStrumMisura: string;
  s8fPotAssorbitaKw: number;
  s8gFlgStrumentazioneFissa: number;
  s8hOperatoreDenominazione: string;
  s8iOperatoreNumIscriz: string;
  s8jSurriscaldamentoK: number;
  s8jTempSorgIngressoC: number;
  s8jSottoraffreddamentoK: number;
  s8jTempSorgUscitaC: number;
  s8jTempCondensazioneC: number;
  s8jTempIngressoFluidoC: number;
  s8jTempEvaporazioneC: number;
  s8jTempUscitaFluidoC: number;
  s9aFlgVerificaSuperata: number;
  s9bFlgRispettoNormativa: number;
  s9cFlgNoRispetto7a: number;
  s9cFlgNoRispetto7b: number;
  s9cFlgNoRispetto8d: number;
  s9cFlgNoRispetto9a: number;
}

interface DatiRapProvaWebGf {
  s1cFlgReeInviato: number;
  s5aFlgIsolamCanaliDistrib: number;
  s5aFlgIsolamReteDistrib: number;
  s2eFlgTrattH2oNonRich: number;
  s3dFlgCoibentazioniIdonee: number;
  s3aFlgLocaleIntIdoneo: number;
  s3bFlgLineeElettrIdonee: number;
  s4dFlgLibUsoPresente: number;
  s3cFlgVentilazAdeguate: number;
  s5aFlgSostituzSistemiReg: number;
  s5aFlgSostituzMacchineReg: number;
  s4aFlgLibImpPresente: number;
  s4cFlgConformitaPresente: number;
  s4bFlgLibCompilato: number;
  s1lFlgDelega: number;
  s1cSiglaBollino: string;
  s1lDenomDelegato: string;
  s1ePotTermicaMaxKw: number;
  s1eDtPrimaInstallazione: number;
  s1cDataRee: number;
  s1cNumBollino: number;
  s1cFlgReeBollino: number;
  s1cSiglaBollino2: string;
  s1cSiglaBollino3: string;
  s1cFlgSiglaBollino: string;
  s5bFlgNoIntervConv: number;
  s5bFlgRelazDettaglio: number;
  s5bFlgRelazDettaglioSucc: number;
  s5bFlgValutazNonEseguita: number;
  s5bMotivoRelazNonEseg: string;
  s5cFlgDimensCorretto: number;
  s5cFlgDimensNonCorretto: number;
  s5cFlgDimensNonControll: number;
  s5cFlgDimensRelazSucces: number;
  idDettIspezGf?: number;
  progressivo: number;
  dataInstall: number;
  fOsservazioni: string;
  fRaccomandazioni: string;
  fPrescrizioni: string;
  s6hFlgInverter: number;
  s6nFlgFugaDiretta: number;
  s6nFlgFugaIndiretta: number;
  s7aFkFrequenzaManut: number;
  s7aFrequenzaManutaltro: string;
  s7aFlgManutEffettuata: number;
  s7aDataUltimaManut: number;
  s7bFlgRegistroApparecc: number;
  s7cFlgReePresente: number;
  s7cDataRee: number;
  s7cFlgOsservazioni: number;
  s7cFlgRaccomand: number;
  s7cFlgPrescr: number;
  circuiti: Circuito[];
}
// ...
export interface SaveRapprovaWeb {
  datiRapProva: DatiRapProva;
  datiRapProvaWebGt?: DatiRapProvaWebGt;
  datiRapProvaWebGf?: DatiRapProvaWebGf;
  // listDatiGt?: DatiGt[];
  // listDatiGf?: DatiGf[];
  // listDatiBR: DatiBr[];
}